🚚 #107 :: 자동 삭제 기능 서비스 분리
