+++ src/main/java/com/example/mukgen/domain/mealsuggestion/service/MealSuggestionService.java
@@ -1,22 +1,20 @@
 package com.example.mukgen.domain.mealsuggestion.service;
 
+import com.example.mukgen.domain.mealsuggestion.controller.dto.request.MealSuggestionCreateRequest;
 import com.example.mukgen.domain.mealsuggestion.controller.dto.request.MealSuggestionUpdateRequest;
 import com.example.mukgen.domain.mealsuggestion.controller.dto.response.MealSuggestionResponse;
 import com.example.mukgen.domain.mealsuggestion.entity.MealSuggestion;
 import com.example.mukgen.domain.mealsuggestion.repository.MealSuggestionRepository;
-import com.example.mukgen.domain.mealsuggestion.controller.dto.request.MealSuggestionCreateRequest;
-import com.example.mukgen.domain.user.service.exception.NoPermissionException;
 import com.example.mukgen.domain.mealsuggestion.service.exception.MealSuggestionNotFoundException;
 import com.example.mukgen.domain.mealsuggestion.service.exception.MealSuggestionWriterMissMatchException;
+import com.example.mukgen.domain.user.entity.User;
 import com.example.mukgen.domain.user.entity.type.UserRole;
 import com.example.mukgen.domain.user.service.UserFacade;
-import com.example.mukgen.domain.user.entity.User;
+import com.example.mukgen.domain.user.service.exception.NoPermissionException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 @Service
@@ -67,14 +65,6 @@ public class MealSuggestionService {
             throw MealSuggestionWriterMissMatchException.EXCEPTION;
 
         mealSuggestionRepository.delete(mealSuggestion);
-    }
-
-    @Scheduled(cron = "0 0 0 * *")
-    @Transactional
-    public void autoRemoveMealSuggestion() {
-        mealSuggestionRepository
-                .removeByCreatedAtEquals(
-                        LocalDateTime.now().minusWeeks(1));
     }
 
     public List<MealSuggestionResponse> findAllSuggestion(
